Start adding the cold warrior.
A deeply outdated person, who lived under the spectre of NT going to interstellar war. The role requires being overly paranoid and doomsday prepping. They should be expected to gather weapons, build a bunker, and prepare for the end of the world. Ultimately, producing a shuttle and living just off station may be called for.